Add datatype conversion to createIntakeBodyValidator

+++ root/backend/app/validators/intake-validator.ts
@@ -25,7 +25,8 @@ export const createIntakeBodyValidator: any = [
   body('intakeId')
     .notEmpty().withMessage('Missing intakeId')
     .isInt({ min: 200001, max: 209912 }) // covers years 2000-2099, months 01-12
-    .withMessage("Invalid intakeId format. Must be an integer in YYYYMM format (e.g., 202503, 202508, 202408)."),
+    .withMessage("Invalid intakeId format. Must be an integer in YYYYMM format (e.g., 202503, 202508, 202408).")
+    .toInt(),
   validate,
 ];
